(refactor) package structure change
