+++ src/test/java/ladder/domain/LadderTest.java
@@ -1,9 +1,11 @@
-package ladder;
+package ladder.domain;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.junit.jupiter.api.Assertions.assertAll;
 
 import java.util.List;
+import ladder.domain.Ladder;
+import ladder.domain.Line;
 import org.junit.jupiter.api.Test;
 
 public class LadderTest {
