+++ src/test/java/ladder/domain/NamesTest.java
@@ -1,9 +1,10 @@
-package ladder;
+package ladder.domain;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.junit.jupiter.api.Assertions.assertAll;
 
 import java.util.List;
+import ladder.domain.Names;
 import org.junit.jupiter.api.Test;
 
 public class NamesTest {
