+++ src/test/java/ladder/view/InputViewTest.java
@@ -1,4 +1,4 @@
-package ladder;
+package ladder.view;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
@@ -6,6 +6,7 @@ import java.io.ByteArrayInputStream;
 import java.io.ByteArrayOutputStream;
 import java.io.PrintStream;
 import java.nio.charset.StandardCharsets;
+import ladder.view.InputView;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
